Extension/requirements.txt
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
